corrected code for pimon game
The previous "final" file for this still had some errors in it...I know I made some typos and messed up some indentation...but there are other issues that I think came from the original code. Either way, THIS is the file that should work without you having to mess with stuff.